Groundwork
Takes the skeleton, and implements some of the rules in accordance with the MRState Diagram. Some lines of code are commented out; this is because the relevant classes that are in the class by association are not implemented, and the functionality of the methods within those other classes are not yet known. 
Note: might have to shuffle the work for some people around, as the entities on the right-hand side of the class diagram will probably need to be implemented for ManagementRecord and the Database to be implemented thoroughly.