Logic Fix (OR -> AND)
Say we're going into the method with In_Transit status. It will run through the check like so.
Status (IN_TRANSIT) != IN_TRANSIT == False
Status (IN_TRANSIT) != DEPARTING_LOCAL_AIRSPACE == True

The If check is trying to catch INVALID entries into the method. So the check should only be true if the status is invalid and we should immediately exit out of the method.

In the above example, if an OR is used, you will have (False || True), and it will return true. If an AND is used, you will have (False && True), which will return false. Same if the status is local airspace. Therefore, the two valid statuses will skip the IF check by inserting a FALSE clause into the AND check. Any invalid status will return true for BOTH checks, and therefore will return TRUE for the AND, and will immediately return.