LATC Completion
This should include full functionality for the LATC class. If changes are needed, let me know; I think they should be minor, and related to the fact that only valid aircraft should show in the list. There are some UI display issues that I might return to if there's enough time to do so.