CleaningSupervisor - Working Version
This version of the cleaning supervisor should have a working GUI and functional button, and implemented as an observer of the AMDatabase. I included a border around the JList in this GUI; I think it might look good if all the text field components are bordered, if we have the time to do that.